correct production deployment issues
